# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/matches": {
      "get": {
        "operationId": "MatchesController_listRooms",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "WAITING",
                "ACTIVE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MatchRoomDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "total",
                        "page",
                        "limit",
                        "totalPages"
                      ],
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      },
      "post": {
        "operationId": "MatchesController_createRoom",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/MatchDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
...
    "/api/v1/matches/{matchId}": {
      "get": {
        "operationId": "MatchesController_getMatch",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/MatchDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      },
      "delete": {
        "operationId": "MatchesController_cancelRoom",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/api/v1/matches/{matchId}/participants": {
      "post": {
        "operationId": "MatchesController_joinRoom",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/MatchDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/api/v1/matches/{matchId}/test-inputs": {
      "get": {
        "operationId": "MatchesController_testInputs",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/TestInputsDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
    "/api/v1/matches/{matchId}/submissions": {
      "post": {
        "operationId": "MatchesController_submit",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowserJudgedSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/JudgedSubmissionDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Matches"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coding Arena",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProblemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "timeLimitMs": {
            "type": "number"
          },
          "category": {
            "type": "string",
            "enum": [
              "BASICS",
              "CONDITIONS",
              "LOOPS",
              "STRINGS",
              "LISTS",
              "MATH",
              "ALGORITHMS"
            ]
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "EASY",
              "MEDIUM",
              "HARD"
            ]
          },
          "isBuiltIn": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "timeLimitMs",
          "category",
          "difficulty",
          "isBuiltIn",
          "description"
        ]
      },
...
      "DeletedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
...
      "TestInputsDto": {
        "type": "object",
        "properties": {
          "problemId": {
            "type": "string"
          },
          "timeLimitMs": {
            "type": "number"
          },
          "inputs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "problemId",
          "timeLimitMs",
          "inputs"
        ]
      },
      "JudgedSubmissionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "EVALUATING",
              "ACCEPTED",
              "WRONG_ANSWER",
              "TIME_LIMIT_EXCEEDED",
              "RUNTIME_ERROR",
              "COMPILATION_ERROR"
            ]
          },
          "language": {
            "type": "string",
            "enum": [
              "PYTHON",
              "JAVASCRIPT",
              "TYPESCRIPT"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "evaluatedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "failedTest": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "status",
          "createdAt",
          "evaluatedAt",
          "failedTest"
        ]
      },
      "BrowserJudgedSubmissionDto": {
        "type": "object",
        "properties": {
          "problemId": {
            "type": "string"
          },
          "sourceCode": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "enum": [
              "PYTHON",
              "JAVASCRIPT",
              "TYPESCRIPT"
            ]
          },
          "outcome": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "RUNTIME_ERROR",
              "TIME_LIMIT_EXCEEDED",
              "COMPILATION_ERROR"
            ]
          },
          "outputs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "problemId",
          "sourceCode",
          "language",
          "outcome",
          "outputs"
        ]
      },
      "SubmissionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "EVALUATING",
              "ACCEPTED",
              "WRONG_ANSWER",
              "TIME_LIMIT_EXCEEDED",
              "RUNTIME_ERROR",
              "COMPILATION_ERROR"
            ]
          },
          "language": {
            "type": "string",
            "enum": [
              "PYTHON",
              "JAVASCRIPT",
              "TYPESCRIPT"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "evaluatedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "status",
          "createdAt",
          "evaluatedAt"
        ]
      },
...
      "ParticipantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "eloRating": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "displayName",
          "eloRating"
        ]
      },
      "MatchRoomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "WAITING",
              "ACTIVE"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "playerOne": {
            "$ref": "#/components/schemas/ParticipantDto"
          },
          "playerTwo": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ParticipantDto"
              }
            ]
          },
          "isMine": {
            "type": "boolean"
          },
          "currentRound": {
            "type": "number",
            "nullable": true
          },
          "playerOneWins": {
            "type": "number"
          },
          "playerTwoWins": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "createdAt",
          "playerOne",
          "playerTwo",
          "isMine",
          "currentRound",
          "playerOneWins",
          "playerTwoWins"
        ]
      },
      "MatchRoundDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "roundNumber": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "WON",
              "DRAW"
            ]
          },
          "winnerId": {
            "type": "string",
            "nullable": true
          },
          "endsAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "problem": {
            "$ref": "#/components/schemas/ProblemDto"
          },
          "submissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubmissionDto"
            }
          }
        },
        "required": [
          "id",
          "roundNumber",
          "status",
          "winnerId",
          "endsAt",
          "problem",
          "submissions"
        ]
      },
      "MatchDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "matched"
            ]
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "WAITING",
              "ACTIVE",
              "COMPLETED",
              "DRAW"
            ]
          },
          "winnerId": {
            "type": "string",
            "nullable": true
          },
          "myPlayerId": {
            "type": "string"
          },
          "playerOne": {
            "$ref": "#/components/schemas/ParticipantDto"
          },
          "playerTwo": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ParticipantDto"
              }
            ]
          },
          "currentRound": {
            "type": "number",
            "nullable": true
          },
          "rounds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchRoundDto"
            }
          }
        },
        "required": [
          "state",
          "id",
          "title",
          "status",
          "winnerId",
          "myPlayerId",
          "playerOne",
          "playerTwo",
          "currentRound",
          "rounds"
        ]
      },
      "CreateMatchRoomDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 60
          },
          "problemIds": {
            "minItems": 3,
            "maxItems": 3,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string",
            "enum": [
              "BASICS",
              "CONDITIONS",
              "LOOPS",
              "STRINGS",
              "LISTS",
              "MATH",
              "ALGORITHMS"
            ]
          }
        }
      },
...
    }
  }
}